added forms to code

+++ src/main/java/org/example/hexlet/model/Course.java
@@ -7,19 +7,19 @@ import lombok.ToString;
 @Setter
 @ToString
 public class Course {
-    public Integer id;
+    public Long id;
 
     @ToString.Include
     private String name;
     private String description;
 
-    public Course(Integer id, String name, String description) {
+    public Course(Long id, String name, String description) {
         this.id = id;
         this.name = name;
         this.description = description;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
